Model Context Protocol/MCP Build Rich-Context AI Apps with Anthropic/L5/mcp_project/pyproject.toml
[project]
name = "mcp-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.51.0",
    "arxiv>=2.2.0",
    "mcp>=1.8.0",
    "nest-asyncio>=1.6.0",
    "openai>=1.97.1",
    "python-dotenv>=1.1.0",
]
